Backend/requirements.txt
FastApi
uvicorn[standard]
SQLalchemy
Pydantic-settings
python-dotenv
firebase-admi
email-validetor
